+++ frontend/techpick-extension/src/pages/BookmarkPage.tsx
@@ -6,14 +6,14 @@
 // 5. update 로직 구현.
 
 // import { CurrentTabInfo } from '@/components';
-import { TagInput } from '@/widgets';
+import { TagPicker } from '@/widgets';
 
 export function BookmarkPage() {
   return (
     <div style={{ width: '600px', height: '600px' }}>
       <h1>Bookmark page</h1>
 
-      <TagInput />
+      <TagPicker />
       {/* <CurrentTabInfo /> */}
     </div>
   );
