+++ frontend/techpick-extension/src/widgets/index.ts
@@ -1,1 +1,1 @@
-export * from './TagInput';
+export * from './TagPicker';
